moved conact components into a contact file
