+++ src/index.js
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom/client';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
-import ContactPage from './Components/ContactPage';
+import ContactPage from './Components/Contact/ContactPage';
 
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
